File

Description

Documentation

data File

Constructors

NormalFile
Normal file with text content inside

Fields

name :: String

The name of the file

content :: String

The file contents

Directory
Directory containing other directories or Normal files. The root directory will have empty file name ""

Fields

name :: String

The name of the file

files :: [File]

The files contained in the directory

isRoot

Arguments

:: File
The file that we want to check

-> Bool
Will return true if the current directory is root

Check if the File is the root directory